Fix Windows (#38)

+++ model_api/cpp/models/include/models/results.h
@@ -95,7 +95,7 @@ struct ClassificationResult : public ResultBase {
 };
 
 struct DetectedObject : public cv::Rect2f {
-    unsigned int labelID;
+    size_t labelID;
     std::string label;
     float confidence;
